% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{CSO2001_ANB_3NT2T_Select}
\alias{CSO2001_ANB_3NT2T_Select}
\title{CSO Select Data, long format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25000 rows and 6 columns.}
\usage{
CSO2001_ANB_3NT2T_Select
}
\description{
Table Reference - Tillinghast, "American Council of Life Insurers: Preferred Version of 2001 CSO Mortality Tables",
Towers Perrin (Atlanta, 2006) Attachment A.1.a.i and A.2.a.i Accessed: January, 2013 from
http://www.soa.org/2001-CSO-Preferred-Class-Structure-Mortality-Tables.pdf
Comments - Study Data: A preferred version of the 2001 Commissioners Standard Ordinary (CSO) table
(See SOA Table Identities 1136-1141) was not initially constructed as data on preferred risk mortality
was considered inadequate at the time. In 2005, the American Council of Life Insurers engaged Tillinghast
to develop a version of the 2001 CSO for preferred risks. Tillinghast relied on the initial 1990-95 SOA experience data
underlying the 2001 CSO tables and 2001 Valuation Basic Table (VBT) tables (See SOA Table Identities 1148-1153).
They then supplemented the information with data from the SOA's 1996-2001 individual life mortality experience study,
the 1994 and 2005 SOA Preferred Risk Underwriting surveys, Tillinghast's Old Age Mortality Study (TOAMS),
Steve Cox's 2004 article "Does Preferred Wear Off", and various medical studies for specific health impairments.
Methodology: Both the 2001 and 2001 VBT CSO non-smoker mortality tables were split into super-preferred, preferred,
and residual standard. The respective CSO and VBT smoker tables were split into preferred and residual standard.
Ratios of preferred risk to aggregate risk, as well as the principle of conservation of deaths, were used to accomplish
the splits. Estimates were derived for the level of preferred risk early duration mortality and the persistence of
preferred risk mortality differentials. The effects of underwriting selection were incorporated by looking at specific
health impairments, namely: alcoholism, cholesterol levels, blood pressure and diabetes. The same loading formula for
developing the 2001 CSO from the 2001 VBT was used to construct the preferred risk tables. NOTE: Table is a combination
of "Select and Ultimate" A.1.a.i and "Ultimate" A.2.a.i thus ultimate values start at age 16 rather than 25.
Data Transcription Errors: None. Data Certified: 01/2013
}
\keyword{datasets}
